+++ packages/abap-deploy-config-inquirer/CHANGELOG.md
@@ -1,5 +1,30 @@
 # @sap-ux/abap-deploy-config-inquirer
 
+## 0.0.4
+
+### Patch Changes
+
+-   Updated dependencies [d3dafeb]
+    -   @sap-ux/btp-utils@0.15.2
+    -   @sap-ux/axios-extension@1.16.5
+    -   @sap-ux/fiori-generator-shared@0.3.16
+    -   @sap-ux/system-access@0.5.10
+
+## 0.0.3
+
+### Patch Changes
+
+-   b158244: remove feature toggle usage
+
+## 0.0.2
+
+### Patch Changes
+
+-   3e1a83a: FEAT - Move Guided Help code into a new module @sap-ux/guided-answers-helper
+-   Updated dependencies [3e1a83a]
+    -   @sap-ux/fiori-generator-shared@0.3.15
+    -   @sap-ux/guided-answers-helper@0.0.1
+
 ## 0.0.1
 
 ### Patch Changes