+++ packages/deploy-tooling/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @sap-ux/deploy-tooling
 
+## 0.15.15
+
+### Patch Changes
+
+-   Updated dependencies [d3dafeb]
+    -   @sap-ux/btp-utils@0.15.2
+    -   @sap-ux/axios-extension@1.16.5
+    -   @sap-ux/system-access@0.5.10
+
 ## 0.15.14
 
 ### Patch Changes